Add: Login

+++ siswa/siswa.php
@@ -2,6 +2,11 @@
     require 'functions_siswa.php';
     require '../views/template/navbar.php';
     $siswa = query("SELECT * FROM tb_siswa INNER JOIN tb_kelas USING(id_kelas) ORDER BY nis ASC");
+
+    $nis = $_SESSION['key'];
+    if($_SESSION['leveluser'] == $nis ) {
+        $siswa = query("SELECT * FROM tb_siswa INNER JOIN tb_kelas USING(id_kelas) WHERE nis=$nis ORDER BY nis ASC");
+    }
 ?>
 
 <!DOCTYPE html>
@@ -17,7 +22,9 @@
 
 <body>
     <h1>Data Siswa</h1>
+    <?php if($_SESSION['leveluser'] == 'Admin' ) : ?>
     <a href="create_siswa.php">Tambah Siswa</a>
+    <?php endif; ?>
     <table border="1" cellspacing="0" cellpadding="10">
         <tr>
             <th>NIS</th>
@@ -26,7 +33,9 @@
             <th>Kelas</th>
             <th>Jenis Kelamin</th>
             <th>Alamat</th>
-            <th>Aksi</th>
+             <?php if($_SESSION['leveluser'] == 'Admin' ) : ?>
+                <th>Aksi</th>
+            <?php endif; ?>
         </tr>
 
         <?php foreach($siswa as $row) : ?>
@@ -37,10 +46,12 @@
             <td><?= $row["kelas"]; ?></td>
             <td><?= $row["jenis_kelamin"]; ?></td>
             <td><?= $row["alamat"]; ?></td>
-            <td>
-                <a href="delete_siswa.php?nis=<?= $row["nis"]; ?>">Hapus</a>
-                <a href="update_siswa.php?nis=<?= $row["nis"]; ?>">Edit</a>
-            </td>
+            <?php if($_SESSION['leveluser'] == 'Admin' ) : ?>
+                <td>
+                    <a href="delete_siswa.php?nis=<?= $row["nis"]; ?>">Hapus</a>
+                    <a href="update_siswa.php?nis=<?= $row["nis"]; ?>">Edit</a>
+                    </td>
+            <?php endif; ?>
         </tr>
         <?php endforeach; ?>
     </table>
